[refactor/#224] my-notifications.ts 파일명 camelCase로 변경 및 import 경로 수정
